requirements.txt
streamlit
torch
torchvision
ultralytics
opencv-python
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow
tqdm